requirements.txt
alpaca-py==0.8.2
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.1.0
cycler==0.11.0
DateTime==5.1
fonttools==4.40.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
msgpack==1.0.5
numpy==1.24.3
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
pydantic==1.10.9
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sseclient-py==1.7.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
websockets==10.4
zope.interface==6.0